--- docs/solution.md
@@ -1,1 +1,0 @@
-# Solution Structures
