+++ test/test_solution_node.py
@@ -1,6 +1,6 @@
 import operator
 from unittest import TestCase, main
-from solution.solution_node import SolutionNode
+from structures.solution_node import SolutionNode
 
 
 class TestSolutionNode(TestCase):
@@ -8,7 +8,7 @@ class TestSolutionNode(TestCase):
         node = SolutionNode(operator.add)
         node.left_child = SolutionNode(1)
         node.right_child = SolutionNode(1)
-        self.assertEqual(node.evaluate(), 2)
+        self.assertEqual(node.evaluate({}), 2)
 
 
 if __name__ == "__main__":
