+++ test/test_solution_tree.py
@@ -1,7 +1,7 @@
 import operator
 from unittest import TestCase, main
-from solution.solution_node import SolutionNode
-from solution.solution_tree import SolutionTree
+from structures.solution_node import SolutionNode
+from structures.solution_tree import SolutionTree
 
 
 class TestSolutionTree(TestCase):
@@ -13,7 +13,7 @@ class TestSolutionTree(TestCase):
         tree.root.right_child = SolutionNode(operator.mul)
         tree.root.right_child.left_child = SolutionNode(1)
         tree.root.right_child.right_child = SolutionNode(2)
-        self.assertEqual(tree.evaluate(), 14)
+        self.assertEqual(tree.evaluate({}), 14)
 
 
 if __name__ == "__main__":
